udpate

+++ Shop-Management-System/src/main/java/Request/ProductUpdateRequest.java
@@ -1,12 +1,14 @@
 package Request;
 
+import java.math.BigDecimal;
+
 import Entity.Category;
 import lombok.Data;
 
 @Data
 public class ProductUpdateRequest {
 	private String name;
-	private double price;
+	private BigDecimal price;
 	private Category category;
 	private String brand;
 	private Integer inventory;
